requirements.txt
alembic==1.13.0
anyio==3.6.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.3
cryptography==41.0.7
fastapi==0.79.0
h11==0.13.0
idna==3.3
Mako==1.3.0
MarkupSafe==2.1.3
pyaml==23.9.7
pycparser==2.21
pydantic==1.9.1
PyJWT==2.8.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.2.0
SQLAlchemy==2.0.23
starlette==0.19.1
typing_extensions==4.3.0
urllib3==2.1.0
uvicorn==0.18.2
